requirements.txt
datasets
langchain-community
chromadb
transformers
torch
python-telegram-bot
langchain-huggingface
